hide commands, move two commands to common, fix windows svn diff
