+++ src/commands.ts
@@ -16,9 +16,7 @@ import { Log } from "./commands/log";
 import { OpenChangeBase } from "./commands/openChangeBase";
 import { OpenChangeHead } from "./commands/openChangeHead";
 import { OpenChangePrev } from "./commands/openChangePrev";
-import { OpenDiff } from "./commands/openDiff";
 import { OpenFile } from "./commands/openFile";
-import { OpenFileRemote } from "./commands/openFileRemote";
 import { OpenHeadFile } from "./commands/openHeadFile";
 import { OpenResourceBase } from "./commands/openResourceBase";
 import { OpenResourceHead } from "./commands/openResourceHead";
@@ -84,6 +82,4 @@ export function registerCommands(model: Model, disposables: Disposable[]) {
   disposables.push(new DeleteUnversioned());
   disposables.push(new OpenChangeHead());
   disposables.push(new OpenHeadFile());
-  disposables.push(new OpenDiff());
-  disposables.push(new OpenFileRemote());
 }
