+++ src/historyView/common.ts
@@ -1,11 +1,19 @@
 import { createHash } from "crypto";
 import * as fs from "fs";
 import * as path from "path";
-import { env, TreeItem, Uri, window } from "vscode";
+import {
+  commands,
+  env,
+  TextDocumentShowOptions,
+  TreeItem,
+  Uri,
+  window
+} from "vscode";
 import { ISvnLogEntry, ISvnLogEntryPath } from "../common/types";
 import { configuration } from "../helpers/configuration";
 import { IRemoteRepository } from "../remoteRepository";
 import { SvnRI } from "../svnRI";
+import { dumpSvnFile } from "../tempFiles";
 
 export enum LogTreeItemKind {
   Repo = 1,
@@ -218,4 +226,61 @@ export function getCommitToolTip(commit: ISvnLogEntry): string {
 ${date}
 Revision: ${commit.revision}
 Message: ${commit.msg}`;
+}
+
+export async function openDiff(
+  repo: IRemoteRepository,
+  arg: Uri,
+  r1: string,
+  r2: string
+) {
+  const getUri = async (revision: string): Promise<Uri> => {
+    if (revision === "BASE") {
+      const nm = repo.getPathNormalizer();
+      const ri = nm.parse(arg.toString(true));
+      const localPath = ri.localFullPath;
+      if (localPath === undefined || !fs.existsSync(localPath.path)) {
+        const errorMsg =
+          "BASE revision doesn't exist for " +
+          (localPath ? localPath.path : "remote path");
+        window.showErrorMessage(errorMsg);
+        throw new Error(errorMsg);
+      }
+      return localPath;
+    }
+    let out;
+    try {
+      out = await repo.show(arg, revision);
+    } catch (e) {
+      window.showErrorMessage("Failed to open path");
+      throw e;
+    }
+    return dumpSvnFile(arg, revision, out);
+  };
+  const uri1 = await getUri(r1);
+  const uri2 = await getUri(r2);
+  const opts: TextDocumentShowOptions = {
+    preview: true
+  };
+  const title = `${path.basename(arg.path)} (${r1} : ${r2})`;
+  return commands.executeCommand<void>("vscode.diff", uri1, uri2, title, opts);
+}
+
+export async function openFileRemote(
+  repo: IRemoteRepository,
+  arg: Uri,
+  against: string
+) {
+  let out;
+  try {
+    out = await repo.show(arg, against);
+  } catch {
+    window.showErrorMessage("Failed to open path");
+    return;
+  }
+  const localUri = await dumpSvnFile(arg, against, out);
+  const opts: TextDocumentShowOptions = {
+    preview: true
+  };
+  return commands.executeCommand<void>("vscode.open", localUri, opts);
 }
